requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cfgv==3.3.1
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
identify==2.4.12
nodeenv==1.6.0
platformdirs==2.5.1
pre-commit==2.17.0
PyJWT==2.3.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.13.4
